Rename serializer
