+++ extensions/ql-vscode/src/query-history/store/query-history-store.ts
